Cleaned & Organized Source Files
